% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseKOS_functions.R
\name{GetProjection}
\alias{GetProjection}
\title{Computes projection value.}
\usage{
GetProjection(X, Data, Cat, Dvec = NULL, Kw = NULL, w = rep(1,
  ncol(Data)), Sigma = NULL, Gamma = NULL)
}
\arguments{
\item{X}{(m x p) Matrix of unlabelled data with numeric features. This function computes the projection value of each data point in X.}

\item{Data}{(n x p) Matrix of training data with numeric features. Cannot have missing values.}

\item{Cat}{(n x 1) Vector of class membership. Values must be either 1 or 2.}

\item{Dvec}{(n x 1) Discrimiant coefficients vector. Default set to NULL. The user can supply Dvec, but it is recommended to allow GetProjection to automatically generate it.}

\item{Kw}{(n x n) Weighted Gaussian kernel matrix. Default set to NULL. The user can supply Kw, but it is recommended to allow GetProjection to automatically generate it.}

\item{w}{(p x 1) Vector of weights for each data variable. Each coordinate must lie between -1 and 1. Default value are all 1s.}

\item{Sigma}{Gaussian kernel parameter. Must be > 0. Default set to NULL. Function runs SelectParam if user-supplied value is not given.}

\item{Gamma}{Ridge parameter. Must be > 0, and default set to NULL. Function runs SelectParam if user-supplied value is not given.}
}
\value{
\item{PV}{ Projection value of x.}
}
\description{
Produces the centered projection value of x onto the discrimiant function determined by Dvec.
}
\details{
Produces the centered projection values of every point in X onto the discrimiant function determined by Dvec. Let K(X,x) be the (n x 1) vector with i-th coordinate k(x_i, x) for training data x_1, ..., x_n, and let C be the (n x n) centering matrix C=I-(1/n) 1 1^T . The projection value is P(x)=(K(X,x)^T-(1/n)  1^T K)CA. It is presented in equation (6) of [Lapanowski and Gaynanova, preprint].
}
\examples{
Sigma <- 1.325386 #Set parameter values equal to result of SelectParam.
Gamma <- 0.07531579 #Speeds up example.
GetProjection(X = Data$TestData , 
              Data = Data$TrainData , 
              Cat = Data$CatTrain , 
              Sigma = 1.325386 , 
              Gamma = 0.07531579)
}
\references{
Lapanowski, Alexander F., and Gaynanova, Irina. ``Sparse feature selection in kernel discriminant analysis via optimal scoring'', preprint.
}
