% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseKOS_functions.R
\name{SelectParams}
\alias{SelectParams}
\title{Generates the Gaussian kernel, ridge, and sparsity parameters.}
\usage{
SelectParams(Data, Cat, Epsilon = 1e-05)
}
\arguments{
\item{Data}{(n x p) Matrix of training data with numeric features. Cannot have missing values.}

\item{Cat}{(n x 1) Vector of class membership. Values must be either 1 or 2.}

\item{Epsilon}{Numerical stability constant with default value 1e-05. Must be > 0 and is typically chosen to be small.}
}
\value{
A list of 
\item{Sigma}{ Gaussian kernel parameter.}  
\item{Gamma}{ Ridge Parameter.}
\item{Lambda}{ Sparsity parameter.}
}
\description{
Generates the gaussian kernel, ridge, and sparsity parameters for use in sparse kernel optimal scoring using the methods presented in [Lapanowski and Gaynanova, 2019]. The Gaussian kernel parameter is generated using five-fold cross-validation of the misclassification error rate aross the {.05, .1, .2, .3, .5} quantiles of squared-distances between groups. The ridge parameter is generated using a stabilization technique. The sparsity parameter is generated by five-fold cross-validation over a logarithmic grid of 20 values within the interval [1e-09*c, c]. The constant c= 2max(|b|), where b is the vector of the linear term in a quadratic form. This interval is automatically generated. See [Lapanowski and Gaynanova, 2019] for more details.
}
\references{
[Lapanowski and Gaynanova, 2019] ``Sparse Feature Selection in Kernel Discriminant Analysis Via Optimal Scoring'', to appear
}
