% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseKOS_functions.R
\name{SelectRidge}
\alias{SelectRidge}
\title{Automatic stabilization ridge parameter selection.}
\usage{
SelectRidge(Data, Cat, Sigma, Epsilon = 1e-05)
}
\arguments{
\item{Data}{(n x p) Matrix of training data with numeric features. Cannot have missing values.}

\item{Cat}{(n x 1) Vector of class membership. Values must be either 1 or 2.}

\item{Sigma}{Gaussian kernel parameter. Must be > 0.}

\item{Epsilon}{Numerical stability constant with default value 1e-05. Must be > 0 and is typically chosen to be small.}
}
\value{
\item{Gamma}{Ridge parameter.}
}
\description{
An automatic ridge parameter selection method. Uses the stabilization technique presented in [Lapanowski and Gaynanova, 2019].
}
\references{
[Lapanowski and Gaynanova, 2019] ``Sparse Feature Selection in Kernel Discriminant Analysis Via Optimal Scoring'', to appear
}
