\name{sparseKOS-package}
\alias{sparseKOS-package}
\alias{sparseKOS}
\docType{package}
\title{
\packageTitle{sparseKOS}
}
\description{
\packageDescription{sparseKOS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sparseKOS}
\packageIndices{sparseKOS}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{sparseKOS}

Maintainer: \packageMaintainer{sparseKOS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
