% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{Classify}
\alias{Classify}
\title{Classifies data.}
\usage{
Classify(TrainData, TrainCat, Data)
}
\arguments{
\item{TrainData}{(n x p) Matrix of training data with numeric features. Cannot have missing values.}

\item{Data}{(m x p) Matrix of unlabelled data to be classified. Cannot have missing values.}

\item{Cat}{(n x 1) Vector of class membership. Values must be either 1 or 2.}
}
\value{
\item{Predictions}{ (m x 1) Vector of predicted class labels for the data points in Data.}
}
\description{
Returns a (m x 1) vector of predicted group classifications for each data point in Data.
}
\references{
[Lapanowski and Gaynanova, 2019] ``Sparse Feature Selection in Kernel Discriminant Analysis Via Optimal Scoring'', preprint
}
