% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseKOS_functions.R
\name{GetProjection}
\alias{GetProjection}
\title{Computes projection value.}
\usage{
GetProjection(x, Data, Cat, Dvec = NULL, K = NULL, w = rep(1, ncol(Data)),
  Sigma = NULL)
}
\arguments{
\item{x}{(p x 1) Data point which we compute the projection value of.}

\item{Data}{(n x p) Matrix of training data with numeric features. Cannot have missing values.}

\item{Cat}{(n x 1) Vector of class membership. Values must be either 1 or 2.}

\item{Dvec}{(n x 1) Discrimiant coefficients vector.}

\item{K}{(n x n) Kernel matrix.}

\item{w}{(p x 1) Vector of weights for each data variable. Each coordinate must lie between -1 and 1. Default value are all 1s.}

\item{Sigma}{Gaussian kernel parameter. Must be > 0.}
}
\value{
\item{PV}{ Projection value of x.}
}
\description{
Produces the centered projection value of x onto the discrimiant function determined by A.
}
\details{
Produces the centered projection value of x onto the discrimiant function determined by A. Let K(X,x) be the (n x 1) vector with i-th coordinate k(x_i, x) for training data x_1, ..., x_n, and let C be the (n x n) centering matrix C=I-(1/n) 1 1^T . The projection value is P(x)=(K(X,x)^T-(1/n)  1^T K)CA. It is presented in equation (6) of [Lapanowski and Gaynanova, 2019].
}
\references{
[Lapanowski and Gaynanova, 2019] ``Sparse Feature Selection in Kernel Discriminant Analysis Via Optimal Scoring'', to appear
}
