% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseKOS_functions.R
\name{SelectParams}
\alias{SelectParams}
\title{Generates parameters.}
\usage{
SelectParams(Data, Cat, Sigma = NULL, Gamma = NULL, Epsilon = 1e-05)
}
\arguments{
\item{Data}{(n x p) Matrix of training data with numeric features. Cannot have missing values.}

\item{Cat}{(n x 1) Vector of class membership. Values must be either 1 or 2.}

\item{Epsilon}{Numerical stability constant with default value 1e-05. Must be > 0 and is typically chosen to be small.}
}
\value{
A list of 
\item{Sigma}{ Gaussian kernel parameter.}  
\item{Gamma}{ Ridge Parameter.}
\item{Lambda}{ Sparsity parameter.}
}
\description{
Generates parameters to be used in sparse kernel optimal scoring.
}
\details{
Generates the gaussian kernel, ridge, and sparsity parameters for use in sparse kernel optimal scoring using the methods presented in [Lapanowski and Gaynanova, preprint]. 
The Gaussian kernel parameter is generated using five-fold cross-validation of the misclassification error rate aross the {.05, .1, .2, .3, .5} quantiles of squared-distances between groups. 
The ridge parameter is generated using a stabilization technique developed in [Lapanowski and Gaynanova, preprint].
The sparsity parameter is generated by five-fold cross-validation over a logarithmic grid of 20 values in an automatically-generated interval.
}
\examples{
Parameters <- SelectParams(Data = Data$TrainData , Cat = Data$CatTrain)
print(Parameters)
}
\references{
Lapanowski, Alexander F., and Gaynanova, Irina. ``Sparse feature selection in kernel discriminant analysis via optimal scoring'', (preprint)

Lancewicki, Tomer. "Regularization of the kernel matrix via covariance matrix shrinkage estimation." arXiv preprint arXiv:1707.06156 (2017).
}
