% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseKOS_functions.R
\name{SparseKernOptScore}
\alias{SparseKernOptScore}
\title{Sparse kernel optimal scoring}
\usage{
SparseKernOptScore(Data, Cat, w0 = rep(1, ncol(Data)), Lambda, Gamma, Sigma,
  Maxniter = 100, Epsilon = 1e-05, Error = 1e-05)
}
\arguments{
\item{Data}{(n x p) Matrix of training data with numeric features. Cannot have missing values.}

\item{Cat}{(n x 1) Vector of class membership. Values must be either 1 or 2.}

\item{w0}{(p x 1) Vector of initial weights for each data variable. Each coordinate must lie between -1 and 1. Default value are all 1s.}

\item{Lambda}{Scalar sparsity parameter on weight vector. Must be >= 0. When Lambda = 0, SparseKOS defaults to kernel optimal scoring of [Lapanowski and Gaynanova, preprint] without sparse feature selection.}

\item{Gamma}{Scalar ridge parameter used in kernel optimal scoring. Must be > 0.}

\item{Sigma}{Scalar Gaussian kernel parameter. Must be > 0.}

\item{Maxniter}{Maximum number of iterations allowed. Default value is 100.}

\item{Epsilon}{Numerical stability constant with default value 1e-05. Must be > 0, and is typically chosen to be small.}

\item{Error}{Scalar which determines convergence of sparse kernel optimal scoring.}
}
\value{
A list of
 \item{Dvec}{ (n x 1) Discrimiant coefficients vector.}
 \item{Weights}{ (p x 1) Final weight vector.}
}
\description{
Implementation of sparse kernel optimal scoring from [Lapanowski and Gaynanova, preprint].
}
\details{
A non-linear binary classifier with simultaneous sparse feature selection. Alternates between solving a kernel ridge regression problem within an optimal scoring framework and solving a Lasso problem on the data features.
Uses the Gaussian kernel. 
The algorithm has three parameters: a kernel, ridge, and sparsity parameter with specifications detailed in the parameter documentation.
}
\examples{
Sigma <- 1.325386  #Set parameter values equal to result of SelectParam.
Gamma <- 0.07531579 
Lambda <- 0.002855275
output <- SparseKernOptScore(Data = Data$TrainData,
                             Cat = Data$CatTrain,
                             Lambda = Lambda,
                             Gamma = Gamma,
                             Sigma = Sigma)
print(output)
}
\references{
Lapanowski, Alexander F., and Gaynanova, Irina. ``Sparse feature selection in kernel discriminant analysis via optimal scoring'', preprint.
}
